moved you died text to feed
